QC Partition

+++ Core/Inc/draw.h
@@ -28,7 +28,7 @@ void draw_qc_menu();
 void draw_qc_menu_deselect(uint16_t pos);
 void draw_qc_menu_focus(uint16_t pos);
 void draw_main_menu_selection(uint16_t move, uint16_t previous_move);
-void draw_power_menu(float volts, uint32_t amperage, uint32_t power);
+void draw_power_menu(uint32_t volts, uint32_t amperage, uint32_t power);
 void draw_current_control_menu(int amperage_load);
 void graph_builder(int value, int lower_bound, int upper_bound);
 void draw_graph_builder_menu(int lower_bound, int upper_bound, const graph_t *graphs, int curr_graph);
